added some client info

+++ src/main/java/com/siewe/inventorymanagementsystem/controller/UserController.java
@@ -117,6 +117,17 @@ public class UserController {
         return new ResponseEntity<UserDto>(result, HttpStatus.OK);
     }
 
+    @GetMapping("/all-users")
+    public Page<UserDto> getAllUsers(@RequestParam(name = "page", defaultValue = "0") Integer page,
+                                     @RequestParam(name = "size", defaultValue = "5") Integer size,
+                                     @RequestParam(name = "sortBy", defaultValue = "createdDate") String sortBy,
+                                     @RequestParam(name = "direction", defaultValue = "desc") String direction
+    ) {
+        log.debug("REST request to get Users");
+        return userService.findAll(page, size, sortBy, direction);
+    }
+
+
     /**
      * GET  /users : get all the users by role.
      *
@@ -181,8 +192,11 @@ public class UserController {
     @DeleteMapping("/users/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable Long id) {
         log.debug("REST request to delete User : {}", id);
+        if (id == null) {
+            return new ResponseEntity<>(new CustomErrorType("Unable to delete. A user id can not be null."), HttpStatus.BAD_REQUEST);
+        }
         userService.delete(id);
-        return new ResponseEntity<User>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
